backend/requirements.txt
fastapi
pydantic
sqlalchemy
openai
python-dotenv
uvicorn
